+++ src/main/java/br/com/marvel/domain/dto/CharacterDataWrapperResponse.java
@@ -1,0 +1,9 @@
+package br.com.marvel.domain.dto;
+
+import java.io.Serializable;
+
+public class CharacterDataWrapperResponse implements Serializable {
+
+	private static final long serialVersionUID = -2548461049044162639L;
+
+}
