+++ src/main/java/br/com/marvel/domain/model/ComicList.java
@@ -1,4 +1,4 @@
-package br.com.marvel.model;
+package br.com.marvel.domain.model;
 
 import java.io.Serializable;
 import java.util.List;
@@ -16,31 +16,31 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "event_list")
-public class EventList implements Serializable {
+@Table(name = "tb_comic_list")
+public class ComicList implements Serializable {
 
-	private static final long serialVersionUID = 4253080937560418455L;
+	private static final long serialVersionUID = -2264303531840109055L;
 
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	//The number of total available events in this list. Will always be greater than or equal to the "returned" value.,
+	//The number of total available issues in this list. Will always be greater than or equal to the "returned" value.,
 	@Column(name = "total_available")
 	private int available;
 
-	//The number of events returned in this collection (up to 20).,
+	//The number of issues returned in this collection (up to 20).,
 	@Column(name = "returned")
 	private int returned;
 
-	//The path to the full list of events in this collection.,
+	//The path to the full list of issues in this collection.,
 	@Column(name = "collection_uri")
 	private String collectionUri;
 
-	//The list of returned series in this collection.
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "eventList")
-	private List<EventSummary> listEventSummary;
+	//The list of returned issues in this collection.
+	@OneToMany(fetch = FetchType.LAZY, mappedBy = "comicList")
+	private List<ComicSummary> listComicSummary;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "character_list_id")
@@ -78,12 +78,12 @@ public class EventList implements Serializable {
 		this.collectionUri = collectionUri;
 	}
 
-	public List<EventSummary> getListEventSummary() {
-		return listEventSummary;
+	public List<ComicSummary> getListComicSummary() {
+		return listComicSummary;
 	}
 
-	public void setListEventSummary(List<EventSummary> listEventSummary) {
-		this.listEventSummary = listEventSummary;
+	public void setListComicSummary(List<ComicSummary> listComicSummary) {
+		this.listComicSummary = listComicSummary;
 	}
 
 	public Character getCharacterList() {
@@ -104,10 +104,10 @@ public class EventList implements Serializable {
 		if (this == obj) {
 			return true;
 		}
-		if (!(obj instanceof EventList)) {
+		if (!(obj instanceof ComicList)) {
 			return false;
 		}
-		EventList other = (EventList) obj;
+		ComicList other = (ComicList) obj;
 		return Objects.equals(id, other.id);
 	}
 }
