+++ src/main/java/br/com/marvel/domain/model/StoryList.java
@@ -1,4 +1,4 @@
-package br.com.marvel.model;
+package br.com.marvel.domain.model;
 
 import java.io.Serializable;
 import java.util.List;
@@ -16,31 +16,31 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "comic_list")
-public class ComicList implements Serializable {
+@Table(name = "tb_story_list")
+public class StoryList implements Serializable {
 
-	private static final long serialVersionUID = -2264303531840109055L;
+	private static final long serialVersionUID = -3009376004251360355L;
 
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	//The number of total available issues in this list. Will always be greater than or equal to the "returned" value.,
+	//The number of total available stories in this list. Will always be greater than or equal to the "returned" value.,
 	@Column(name = "total_available")
 	private int available;
 
-	//The number of issues returned in this collection (up to 20).,
+	//The number of stories returned in this collection (up to 20).,
 	@Column(name = "returned")
 	private int returned;
 
-	//The path to the full list of issues in this collection.,
+	//The path to the full list of stories in this collection.,
 	@Column(name = "collection_uri")
 	private String collectionUri;
 
-	//The list of returned issues in this collection.
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "comicList")
-	private List<ComicSummary> listComicSummary;
+	//The list of returned stories in this collection.
+	@OneToMany(fetch = FetchType.LAZY, mappedBy = "storyList")
+	private List<StorySummary> listStorySummary;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "character_list_id")
@@ -78,12 +78,12 @@ public class ComicList implements Serializable {
 		this.collectionUri = collectionUri;
 	}
 
-	public List<ComicSummary> getListComicSummary() {
-		return listComicSummary;
+	public List<StorySummary> getListStorySummary() {
+		return listStorySummary;
 	}
 
-	public void setListComicSummary(List<ComicSummary> listComicSummary) {
-		this.listComicSummary = listComicSummary;
+	public void setListStorySummary(List<StorySummary> listStorySummary) {
+		this.listStorySummary = listStorySummary;
 	}
 
 	public Character getCharacterList() {
@@ -104,10 +104,10 @@ public class ComicList implements Serializable {
 		if (this == obj) {
 			return true;
 		}
-		if (!(obj instanceof ComicList)) {
+		if (!(obj instanceof StoryList)) {
 			return false;
 		}
-		ComicList other = (ComicList) obj;
+		StoryList other = (StoryList) obj;
 		return Objects.equals(id, other.id);
 	}
 }
