+++ src/main/java/br/com/marvel/domain/model/Url.java
@@ -1,4 +1,4 @@
-package br.com.marvel.model;
+package br.com.marvel.domain.model;
 
 import java.io.Serializable;
 import java.util.Objects;
@@ -14,7 +14,7 @@ import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "url")
+@Table(name = "tb_url")
 public class Url implements Serializable {
 
 	private static final long serialVersionUID = -296617366405646919L;
