+++ src/main/resources/application.properties
@@ -1,5 +1,7 @@
-
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
+requireSSL=false&useSSL=false
+logging.level.org.hibernate.SQL= DEBUG
 spring.datasource.url=jdbc:mysql://localhost:3306/marvel
 spring.datasource.username=root
 spring.datasource.password=root
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL55Dialect
